fix(regrefugee): fixed refugee registration screen
Now, the screen automatically sets state before getting refugee's email from AsyncStorage and checks
if the current user is the primary, avoiding to send '' (empty string) as an email to the database
or the register a secondary member with an email from the primary. Also fixed routing back when
secondary contact registration is completed.